Mergeable version.

+++ includes/render.h
@@ -6,7 +6,7 @@
 /*   By: gaerhard <gaerhard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/09 13:20:37 by lnicosia          #+#    #+#             */
-/*   Updated: 2019/07/08 15:11:23 by lnicosia         ###   ########.fr       */
+/*   Updated: 2019/07/09 11:57:00 by lnicosia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,20 +80,20 @@ typedef struct		s_render
 	int				currentx;
 	double			floor1;
 	double			floor2;
-	int				current_floor;
-	int				current_ceiling;
-	int				current_neighbor_floor;
-	int				current_neighbor_ceiling;
+	double			current_floor;
+	double			current_ceiling;
+	double			current_neighbor_floor;
+	double			current_neighbor_ceiling;
 	double			max_floor;
 	double			max_ceiling;
-	int				max_neighbor_floor;
-	int				max_neighbor_ceiling;
+	double			max_neighbor_floor;
+	double			max_neighbor_ceiling;
 	double			ceiling1;
 	double			ceiling2;
-	int				neighbor_floor1;
-	int				neighbor_floor2;
-	int				neighbor_ceiling1;
-	int				neighbor_ceiling2;
+	double			neighbor_floor1;
+	double			neighbor_floor2;
+	double			neighbor_ceiling1;
+	double			neighbor_ceiling2;
 	int				preclip_x1;
 	int				preclip_x2;
 	int				preclip_floor1;
